backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alpaca-py==0.42.1
alpaca-trade-api==3.2.0
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.1.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
chromedriver-py==133.0.6943.141
decouple==0.0.7
deprecation==2.1.0
Django==5.1.6
django-cors-headers==4.7.0
dotenv==0.9.9
finnhub-python==2.4.24
frozendict==2.4.6
frozenlist==1.7.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.14.0
idna==3.10
msgpack==1.0.3
multidict==6.6.4
multitasking==0.0.11
numpy==2.2.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
peewee==3.17.9
platformdirs==4.3.6
propcache==0.3.2
psycopg==3.2.5
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.1
requests==2.32.3
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
sseclient-py==1.8.0
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==10.4
wsproto==1.2.0
yarl==1.20.1
yfinance==0.2.54
